backend/requirements.txt
# fastapi
# uvicorn[standard]
# pydantic-settings
# python-dotenv
# python-jose[cryptography]
# passlib[bcrypt]
# sqlalchemy
# alembic
# psycopg2-binary   # for PostgreSQL later (safe to keep even for SQLite)


fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
email-validator==2.1.0